// src/common/enums/logs.enum.ts
export enum LOGS_ENUM {


    FILE_UPLOAD_CHUNK_STARTED = 'FILE_UPLOAD_CHUNK_STARTED',
    FILE_UPLOAD_CHUNK_SINGLE_SAVED = 'FILE_UPLOAD_CHUNK_SINGLE_SAVED',
    FILE_UPLOAD_CHUNK_MERGE_STARTED = 'FILE_UPLOAD_CHUNK_MERGE_STARTED',
    FILE_UPLOAD_CHUNK_MERGE_SUCCESS = 'FILE_UPLOAD_CHUNK_MERGE_SUCCESS',
    FILE_UPLOAD_CHUNK_SUCCESS = 'FILE_UPLOAD_CHUNK_SUCCESS',
    FILE_UPLOAD_CHUNK_ERROR = 'FILE_UPLOAD_CHUNK_ERROR',
    FILE_UPLOAD_INSIGHT_METADATA_NOT_FOUND = 'FILE_UPLOAD_INSIGHT_METADATA_NOT_FOUND',
    FILE_UPLOAD_ALREADY_UPLOADED = 'FILE_UPLOAD_ALREADY_UPLOADED',

    FILE_INACTIVE_STARTED = 'FILE_INACTIVE_STARTED',
    FILE_INACTIVE_SUCCESS = 'FILE_INACTIVE_SUCCESS',
    FILE_INACTIVE_ERROR = 'FILE_INACTIVE_ERROR',
    FILE_INACTIVE_GENAI_API_CALLED = 'FILE_INACTIVE_GENAI_API_CALLED',
    FILE_INACTIVE_REQUEST_NOT_FOUND = 'FILE_INACTIVE_REQUEST_NOT_FOUND',
    FILE_INACTIVE_ALREADY_INACTIVE = 'FILE_INACTIVE_ALREADY_INACTIVE',

    FILE_DELETE_STARTED = 'FILE_DELETE_STARTED',
    FILE_DELETE_REQUEST_NOT_FOUND = 'FILE_DELETE_REQUEST_NOT_FOUND',
    FILE_DELETE_SUCCESS = 'FILE_DELETE_SUCCESS',

    GENAI_PROCESS_STARTED = 'GENAI_PROCESS_STARTED',
    GENAI_PROCESS_ALREADY_PROCESSED = 'GENAI_PROCESS_ALREADY_PROCESSED',
    GENAI_PROCESS_DATA_NOT_SUFFIECIENT = 'GENAI_PROCESS_DATA_NOT_SUFFIECIENT',
    GENAI_PROCESS_INITAL_RECORD_SAVED = 'GENAI_PROCESS_INITAL_RECORD_SAVED',
    GENAI_PROCESS_API_CALL_SENT = 'GENAI_PROCESS_API_CALL_SENT',
    GENAI_PROCESS_API_CALL_SUCCESS = 'GENAI_PROCESS_API_CALL_SUCCESS',
    GENAI_PROCESS_API_CALL_ERROR = 'GENAI_PROCESS_API_CALL_ERROR',
    GENAI_PROCESS_FAIL = 'GENAI_PROCESS_FAIL',
    GENAI_PROCESS_SUCCESS = 'GENAI_PROCESS_SUCCESS',
    GENAI_PROCESS_ERROR = 'GENAI_PROCESS_ERROR',

    GENAI_WEBHOOK_STARTED = 'GENAI_WEBHOOK_STARTED',
    GENAI_WEBHOOK_ALREADY_SUCCESS = 'GENAI_WEBHOOK_ALREADY_SUCCESS',
    GENAI_WEBHOOK_REQUEST_NOT_FOUND = 'GENAI_WEBHOOK_REQUEST_NOT_FOUND',
    GENAI_WEBHOOK_SUCCESS = 'GENAI_WEBHOOK_SUCCESS',
    GENAI_WEBHOOK_ERROR = 'GENAI_WEBHOOK_ERROR',

}